requirements.txt
python-telegram-bot==20.7
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.0
httpx~=0.25.2
pydantic==2.6.1
nest-asyncio 